Update to v1.6.7 with Phase 9 (LLM Memory) and refactor settings and feedback components
- Update AGENTS.md and CHANGELOG.md to reflect version 1.6.7 and Phase 9 completion
- Add batch query functions _build_entries_out and _build_comments_out in backend feedback router to optimize database queries
- Refactor feedback API endpoints to use batch query functions for entries and comments
- Remove unused imports and variables in frontend admin feedback and user pages
- Add error handling and retry UI in dashboard and lesson pages for data loading and submission errors
- Replace inline user mapping with mapUser utility in multiple frontend auth-related pages
- Replace repeated locale arrays with a single SUPPORTED_LOCALES export and import it in middleware and i18n request
- Add new shared API types for exercises and quota status in frontend/src/types/api.ts
- Refactor frontend settings page into multiple components: ProfileSection, ConversationSection, VoiceSection, UsageLimitsSection, AppearanceSection, BillingSection
- Move avatar upload, profile editing, conversation settings, voice preview, usage limits, appearance, and billing management into respective components
- Add useLogout hook for consistent logout handling
- Replace inline background styles with reusable bg-dot-grid utility class and apply it across multiple pages and layouts
- Add AppLoading component as Suspense fallback for (app) route group
- Add conversation starters list to translations and use localized starters in ConversationMode component
- Add TTS latency metrics logging only in development mode in AudioPlayer component
- Clean up and unify imports and types across frontend codebase
