+++ frontend/src/app/(app)/admin/feedback/page.tsx
@@ -112,7 +112,7 @@ export default function AdminFeedbackPage() {
         setLoading(false)
       }
     },
-    [t, tAdmin]
+    [t]
   )
 
   useEffect(() => {
