+++ frontend/src/app/(auth)/billing/canceled/page.tsx
@@ -7,14 +7,7 @@ export default function BillingCanceledPage() {
   const t = useTranslations('billing')
 
   return (
-    <div
-      className="bg-fl-bg flex min-h-screen items-center justify-center px-4"
-      style={{
-        backgroundImage:
-          'radial-gradient(circle, var(--fl-dot) 1px, transparent 1px)',
-        backgroundSize: '24px 24px',
-      }}
-    >
+    <div className="bg-fl-bg bg-dot-grid flex min-h-screen items-center justify-center px-4">
       <div className="border-fl-border bg-fl-surface w-full max-w-sm space-y-5 border p-8 text-center">
         <div className="text-fl-muted-2 text-2xl">△</div>
         <p className="text-fl-label text-fl-muted-2 font-mono tracking-widest uppercase">
