+++ frontend/src/app/(auth)/reset-password/page.tsx
@@ -49,14 +49,7 @@ function ResetPasswordContent() {
   }
 
   return (
-    <div
-      className="bg-fl-bg flex min-h-screen items-center justify-center px-4"
-      style={{
-        backgroundImage:
-          'radial-gradient(circle, var(--fl-dot) 1px, transparent 1px)',
-        backgroundSize: '24px 24px',
-      }}
-    >
+    <div className="bg-fl-bg bg-dot-grid flex min-h-screen items-center justify-center px-4">
       <div className="w-full max-w-sm">
         <div className="mb-10 flex flex-col items-center">
           <h1 className="text-fl-fg font-mono text-xl font-bold tracking-widest uppercase">
