+++ frontend/src/app/page.tsx
@@ -82,14 +82,7 @@ export default async function Home() {
   }
 
   return (
-    <div
-      className="bg-fl-bg text-fl-fg flex min-h-screen flex-col"
-      style={{
-        backgroundImage:
-          'radial-gradient(circle, var(--fl-dot) 1px, transparent 1px)',
-        backgroundSize: '24px 24px',
-      }}
-    >
+    <div className="bg-fl-bg bg-dot-grid text-fl-fg flex min-h-screen flex-col">
       <script
         type="application/ld+json"
         dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }}
