+++ frontend/src/components/settings/AppearanceSection.tsx
@@ -1,0 +1,54 @@
+'use client'
+
+import { useTranslations } from 'next-intl'
+import { useThemeStore } from '@/store/theme'
+
+export function AppearanceSection() {
+  const t = useTranslations('settings')
+  const theme = useThemeStore((s) => s.theme)
+  const setTheme = useThemeStore((s) => s.setTheme)
+
+  return (
+    <div className="border-fl-border bg-fl-surface mt-4 border p-6">
+      <div className="border-fl-border mb-5 flex items-center gap-2 border-b pb-4">
+        <span className="text-fl-label text-fl-muted-2">●</span>
+        <span className="text-fl-label text-fl-muted-2 font-mono tracking-widest uppercase">
+          {t('sectionAppearance')}
+        </span>
+      </div>
+      <div className="flex items-center justify-between gap-4">
+        <div>
+          <p className="text-fl-fg font-mono text-xs tracking-wide">
+            {t('theme')}
+          </p>
+          <p className="text-fl-label text-fl-muted-2 mt-0.5 font-mono">
+            {theme === 'dark'
+              ? t('darkActive')
+              : theme === 'light'
+                ? t('lightActive')
+                : t('systemActive')}
+          </p>
+        </div>
+        <div className="flex gap-1">
+          {(['system', 'dark', 'light'] as const).map((opt) => (
+            <button
+              key={opt}
+              onClick={() => setTheme(opt)}
+              className={`text-fl-label border px-3 py-2 font-mono tracking-widest uppercase transition-colors ${
+                theme === opt
+                  ? 'border-fl-border-2 text-fl-fg bg-fl-surface-2'
+                  : 'border-fl-border text-fl-muted-2 hover:border-fl-border-2 hover:text-fl-fg'
+              }`}
+            >
+              {opt === 'system'
+                ? t('themeSystem')
+                : opt === 'dark'
+                  ? t('themeDark')
+                  : t('themeLight')}
+            </button>
+          ))}
+        </div>
+      </div>
+    </div>
+  )
+}
