+++ frontend/src/components/whats-new/WhatsNew.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect, useCallback, useMemo } from 'react'
 import { useTranslations, useMessages } from 'next-intl'
 
-const WHATS_NEW_VERSION = 'v1.6.6'
+const WHATS_NEW_VERSION = 'v1.6.7'
 const STORAGE_KEY = `fl_whats_new_seen_${WHATS_NEW_VERSION}`
 const TOUR_KEY = 'fl_tour_done'
