+++ frontend/src/lib/api.ts
@@ -1,5 +1,3 @@
-'use client'
-
 import { useAuthStore } from '@/store/auth'
 import { useLoadingStore } from '@/store/loading'
