+++ messages/de.json
@@ -825,7 +825,39 @@
     "quotaExceededTokens": "Du hast alle monatlichen KI-Token aufgebraucht.",
     "quotaTokens": "Token diesen Monat",
     "memoryUpdated": "Erinnerung gespeichert",
-    "startersHint": "Oder mit einem Thema beginnen"
+    "startersHint": "Oder mit einem Thema beginnen",
+    "starters": [
+      "Kunst und Kreativität",
+      "Bücher und Lesen",
+      "Kindheitserinnerungen",
+      "Stadt vs. Land",
+      "Kochen und Rezepte",
+      "Träume und Ambitionen",
+      "Familie und Beziehungen",
+      "Mode und Stil",
+      "Lieblingsfilme",
+      "Feste und Traditionen",
+      "Essen aus aller Welt",
+      "Gesundheit und Wohlbefinden",
+      "Hobbys und Basteln",
+      "Sprachenlernen",
+      "Neue Fähigkeiten erlernen",
+      "Geld und Budgetplanung",
+      "Musik und Konzerte",
+      "Haustiere und Tiere",
+      "Öffentliche Verkehrsmittel",
+      "Wissenschaft und Weltraum",
+      "Einkaufsgewohnheiten",
+      "Soziale Medien",
+      "Sport und Fitness",
+      "Technologie und Gadgets",
+      "Die Umwelt",
+      "Reiseziele",
+      "TV-Serien",
+      "Wochendpläne",
+      "Arbeit und Karriereziele",
+      "Homeoffice"
+    ]
   },
   "voiceRecorder": {
     "record": "Aufnehmen",
@@ -980,7 +1012,7 @@
   },
   "whatsNew": {
     "title": "Neuigkeiten",
-    "version": "v1.6.6",
+    "version": "v1.6.7",
     "cta": "Verstanden",
     "entry1": {
       "label": "Verbesserungen am Lernplan",
