+++ messages/en.json
@@ -825,7 +825,39 @@
     "quotaExceededTokens": "You have used all your monthly AI tokens.",
     "quotaTokens": "Tokens this month",
     "memoryUpdated": "Memory updated",
-    "startersHint": "Or start with a topic"
+    "startersHint": "Or start with a topic",
+    "starters": [
+      "Art & creativity",
+      "Books & reading",
+      "Childhood memories",
+      "City vs. countryside",
+      "Cooking & recipes",
+      "Dreams & ambitions",
+      "Family & relationships",
+      "Fashion & style",
+      "Favourite films",
+      "Festivals & traditions",
+      "Food from around the world",
+      "Health & wellbeing",
+      "Hobbies & crafts",
+      "Language learning",
+      "Learning new skills",
+      "Money & budgeting",
+      "Music & concerts",
+      "Pets & animals",
+      "Public transport",
+      "Science & space",
+      "Shopping habits",
+      "Social media",
+      "Sports & fitness",
+      "Technology & gadgets",
+      "The environment",
+      "Travel destinations",
+      "TV shows & series",
+      "Weekend plans",
+      "Work & career goals",
+      "Working from home"
+    ]
   },
   "voiceRecorder": {
     "record": "Record",
@@ -980,7 +1012,7 @@
   },
   "whatsNew": {
     "title": "What's New",
-    "version": "v1.6.6",
+    "version": "v1.6.7",
     "cta": "Got it",
     "entry1": {
       "label": "Study plan improvements",
