+++ messages/es.json
@@ -825,7 +825,39 @@
     "quotaExceededTokens": "Has agotado todos tus tokens de IA mensuales.",
     "quotaTokens": "Tokens este mes",
     "memoryUpdated": "Memoria actualizada",
-    "startersHint": "O empieza con un tema"
+    "startersHint": "O empieza con un tema",
+    "starters": [
+      "Arte y creatividad",
+      "Libros y lectura",
+      "Recuerdos de infancia",
+      "Ciudad vs. campo",
+      "Cocina y recetas",
+      "Sueños y ambiciones",
+      "Familia y relaciones",
+      "Moda y estilo",
+      "Películas favoritas",
+      "Fiestas y tradiciones",
+      "Comida de todo el mundo",
+      "Salud y bienestar",
+      "Aficiones y manualidades",
+      "Aprendizaje de idiomas",
+      "Aprender nuevas habilidades",
+      "Dinero y presupuesto",
+      "Música y conciertos",
+      "Mascotas y animales",
+      "Transporte público",
+      "Ciencia y espacio",
+      "Hábitos de compra",
+      "Redes sociales",
+      "Deporte y fitness",
+      "Tecnología y gadgets",
+      "El medio ambiente",
+      "Destinos de viaje",
+      "Series de televisión",
+      "Planes de fin de semana",
+      "Trabajo y objetivos profesionales",
+      "Trabajar desde casa"
+    ]
   },
   "voiceRecorder": {
     "record": "Grabar",
@@ -980,7 +1012,7 @@
   },
   "whatsNew": {
     "title": "Novedades",
-    "version": "v1.6.6",
+    "version": "v1.6.7",
     "cta": "Entendido",
     "entry1": {
       "label": "Mejoras en el plan de estudio",
