+++ messages/fr.json
@@ -825,7 +825,39 @@
     "quotaExceededTokens": "Vous avez utilisé tous vos tokens IA mensuels.",
     "quotaTokens": "Tokens ce mois-ci",
     "memoryUpdated": "Mémoire mise à jour",
-    "startersHint": "Ou commencez par un sujet"
+    "startersHint": "Ou commencez par un sujet",
+    "starters": [
+      "Art et créativité",
+      "Livres et lecture",
+      "Souvenirs d'enfance",
+      "Ville vs. campagne",
+      "Cuisine et recettes",
+      "Rêves et ambitions",
+      "Famille et relations",
+      "Mode et style",
+      "Films préférés",
+      "Fêtes et traditions",
+      "Cuisine du monde entier",
+      "Santé et bien-être",
+      "Loisirs et artisanat",
+      "Apprentissage des langues",
+      "Apprendre de nouvelles compétences",
+      "Argent et budget",
+      "Musique et concerts",
+      "Animaux de compagnie",
+      "Transports en commun",
+      "Sciences et espace",
+      "Habitudes d'achat",
+      "Réseaux sociaux",
+      "Sport et fitness",
+      "Technologie et gadgets",
+      "L'environnement",
+      "Destinations de voyage",
+      "Séries télévisées",
+      "Plans du week-end",
+      "Travail et objectifs de carrière",
+      "Travail à domicile"
+    ]
   },
   "voiceRecorder": {
     "record": "Enregistrer",
@@ -980,7 +1012,7 @@
   },
   "whatsNew": {
     "title": "Nouveautés",
-    "version": "v1.6.6",
+    "version": "v1.6.7",
     "cta": "Compris",
     "entry1": {
       "label": "Améliorations du plan d'étude",
