+++ messages/it.json
@@ -825,7 +825,39 @@
     "quotaExceededTokens": "Hai esaurito tutti i tuoi token IA mensili.",
     "quotaTokens": "Token questo mese",
     "memoryUpdated": "Memoria aggiornata",
-    "startersHint": "O inizia con un argomento"
+    "startersHint": "O inizia con un argomento",
+    "starters": [
+      "Arte e creatività",
+      "Libri e lettura",
+      "Ricordi d'infanzia",
+      "Città vs. campagna",
+      "Cucina e ricette",
+      "Sogni e ambizioni",
+      "Famiglia e relazioni",
+      "Moda e stile",
+      "Film preferiti",
+      "Feste e tradizioni",
+      "Cibo da tutto il mondo",
+      "Salute e benessere",
+      "Hobby e artigianato",
+      "Apprendimento delle lingue",
+      "Imparare nuove competenze",
+      "Soldi e budget",
+      "Musica e concerti",
+      "Animali domestici",
+      "Trasporto pubblico",
+      "Scienza e spazio",
+      "Abitudini di acquisto",
+      "Social media",
+      "Sport e fitness",
+      "Tecnologia e gadget",
+      "L'ambiente",
+      "Destinazioni di viaggio",
+      "Serie TV",
+      "Piani per il weekend",
+      "Lavoro e obiettivi di carriera",
+      "Lavorare da casa"
+    ]
   },
   "voiceRecorder": {
     "record": "Registra",
@@ -980,7 +1012,7 @@
   },
   "whatsNew": {
     "title": "Novità",
-    "version": "v1.6.6",
+    "version": "v1.6.7",
     "cta": "Capito",
     "entry1": {
       "label": "Miglioramenti al piano di studio",
