+++ messages/nl.json
@@ -825,7 +825,39 @@
     "quotaExceededTokens": "Je hebt al je maandelijkse AI-tokens gebruikt.",
     "quotaTokens": "Tokens deze maand",
     "memoryUpdated": "Geheugen bijgewerkt",
-    "startersHint": "Of begin met een onderwerp"
+    "startersHint": "Of begin met een onderwerp",
+    "starters": [
+      "Kunst en creativiteit",
+      "Boeken en lezen",
+      "Kindherinneringen",
+      "Stad vs. platteland",
+      "Koken en recepten",
+      "Dromen en ambities",
+      "Familie en relaties",
+      "Mode en stijl",
+      "Favoriete films",
+      "Festivals en tradities",
+      "Eten van over de hele wereld",
+      "Gezondheid en welzijn",
+      "Hobby's en handwerk",
+      "Talen leren",
+      "Nieuwe vaardigheden leren",
+      "Geld en budgetteren",
+      "Muziek en concerten",
+      "Huisdieren en dieren",
+      "Openbaar vervoer",
+      "Wetenschap en ruimte",
+      "Winkelgewoonten",
+      "Sociale media",
+      "Sport en fitness",
+      "Technologie en gadgets",
+      "Het milieu",
+      "Reisbestemmingen",
+      "Tv-series",
+      "Weekendplannen",
+      "Werk en carrièredoelen",
+      "Thuiswerken"
+    ]
   },
   "voiceRecorder": {
     "record": "Opnemen",
@@ -980,7 +1012,7 @@
   },
   "whatsNew": {
     "title": "Nieuw",
-    "version": "v1.6.6",
+    "version": "v1.6.7",
     "cta": "Begrepen",
     "entry1": {
       "label": "Verbeteringen aan het studieplan",
