+++ messages/pl.json
@@ -825,7 +825,39 @@
     "quotaExceededTokens": "Wykorzystałeś(-aś) wszystkie miesięczne tokeny AI.",
     "quotaTokens": "Tokeny w tym miesiącu",
     "memoryUpdated": "Pamięć zaktualizowana",
-    "startersHint": "Lub zacznij od tematu"
+    "startersHint": "Lub zacznij od tematu",
+    "starters": [
+      "Sztuka i kreatywność",
+      "Książki i czytanie",
+      "Wspomnienia z dzieciństwa",
+      "Miasto vs. wieś",
+      "Gotowanie i przepisy",
+      "Marzenia i ambicje",
+      "Rodzina i relacje",
+      "Moda i styl",
+      "Ulubione filmy",
+      "Festiwale i tradycje",
+      "Jedzenie z całego świata",
+      "Zdrowie i dobre samopoczucie",
+      "Hobby i rękodzieło",
+      "Nauka języków",
+      "Zdobywanie nowych umiejętności",
+      "Pieniądze i budżet",
+      "Muzyka i koncerty",
+      "Zwierzęta domowe",
+      "Transport publiczny",
+      "Nauka i kosmos",
+      "Nawyki zakupowe",
+      "Media społecznościowe",
+      "Sport i fitness",
+      "Technologia i gadżety",
+      "Środowisko",
+      "Miejsca podróży",
+      "Seriale telewizyjne",
+      "Plany na weekend",
+      "Praca i cele zawodowe",
+      "Praca zdalna"
+    ]
   },
   "voiceRecorder": {
     "record": "Nagraj",
@@ -980,7 +1012,7 @@
   },
   "whatsNew": {
     "title": "Co nowego",
-    "version": "v1.6.6",
+    "version": "v1.6.7",
     "cta": "Rozumiem",
     "entry1": {
       "label": "Ulepszenia planu nauki",
