+++ messages/pt.json
@@ -825,7 +825,39 @@
     "quotaExceededTokens": "Você usou todos os seus tokens de IA mensais.",
     "quotaTokens": "Tokens este mês",
     "memoryUpdated": "Memória atualizada",
-    "startersHint": "Ou comece com um tema"
+    "startersHint": "Ou comece com um tema",
+    "starters": [
+      "Arte e criatividade",
+      "Livros e leitura",
+      "Memórias de infância",
+      "Cidade vs. campo",
+      "Culinária e receitas",
+      "Sonhos e ambições",
+      "Família e relacionamentos",
+      "Moda e estilo",
+      "Filmes favoritos",
+      "Festivais e tradições",
+      "Comida do mundo todo",
+      "Saúde e bem-estar",
+      "Hobbies e artesanato",
+      "Aprendizagem de idiomas",
+      "Aprender novas habilidades",
+      "Dinheiro e orçamento",
+      "Música e concertos",
+      "Animais de estimação",
+      "Transporte público",
+      "Ciência e espaço",
+      "Hábitos de compras",
+      "Redes sociais",
+      "Esportes e fitness",
+      "Tecnologia e gadgets",
+      "O meio ambiente",
+      "Destinos de viagem",
+      "Séries de TV",
+      "Planos para o fim de semana",
+      "Trabalho e objetivos de carreira",
+      "Trabalho em casa"
+    ]
   },
   "voiceRecorder": {
     "record": "Gravar",
@@ -980,7 +1012,7 @@
   },
   "whatsNew": {
     "title": "Novidades",
-    "version": "v1.6.6",
+    "version": "v1.6.7",
     "cta": "Entendido",
     "entry1": {
       "label": "Melhorias no plano de estudos",
