+++ messages/ro.json
@@ -825,7 +825,39 @@
     "quotaExceededTokens": "Ai folosit toți tokenii IA lunari.",
     "quotaTokens": "Tokeni luna aceasta",
     "memoryUpdated": "Memorie actualizată",
-    "startersHint": "Sau începe cu un subiect"
+    "startersHint": "Sau începe cu un subiect",
+    "starters": [
+      "Artă și creativitate",
+      "Cărți și lectură",
+      "Amintiri din copilărie",
+      "Oraș vs. sat",
+      "Gătit și rețete",
+      "Vise și ambiții",
+      "Familie și relații",
+      "Modă și stil",
+      "Filme preferate",
+      "Festivaluri și tradiții",
+      "Mâncare din toată lumea",
+      "Sănătate și bunăstare",
+      "Hobby-uri și meșteșuguri",
+      "Învățarea limbilor",
+      "Dobândirea de noi abilități",
+      "Bani și buget",
+      "Muzică și concerte",
+      "Animale de companie",
+      "Transport public",
+      "Știință și spațiu",
+      "Obiceiuri de cumpărături",
+      "Rețele sociale",
+      "Sport și fitness",
+      "Tehnologie și gadgeturi",
+      "Mediul înconjurător",
+      "Destinații de călătorie",
+      "Seriale TV",
+      "Planuri de weekend",
+      "Muncă și obiective de carieră",
+      "Munca de acasă"
+    ]
   },
   "voiceRecorder": {
     "record": "Înregistrează",
@@ -980,7 +1012,7 @@
   },
   "whatsNew": {
     "title": "Noutăți",
-    "version": "v1.6.6",
+    "version": "v1.6.7",
     "cta": "Am înțeles",
     "entry1": {
       "label": "Îmbunătățiri la planul de studiu",
