+++ messages/ru.json
@@ -825,7 +825,39 @@
     "quotaExceededTokens": "Вы использовали все ежемесячные токены ИИ.",
     "quotaTokens": "Токены в этом месяце",
     "memoryUpdated": "Память обновлена",
-    "startersHint": "Или начните с темы"
+    "startersHint": "Или начните с темы",
+    "starters": [
+      "Искусство и творчество",
+      "Книги и чтение",
+      "Воспоминания детства",
+      "Город vs. деревня",
+      "Кулинария и рецепты",
+      "Мечты и амбиции",
+      "Семья и отношения",
+      "Мода и стиль",
+      "Любимые фильмы",
+      "Праздники и традиции",
+      "Еда со всего мира",
+      "Здоровье и благополучие",
+      "Хобби и рукоделие",
+      "Изучение языков",
+      "Освоение новых навыков",
+      "Деньги и бюджет",
+      "Музыка и концерты",
+      "Домашние животные",
+      "Общественный транспорт",
+      "Наука и космос",
+      "Привычки шопинга",
+      "Социальные сети",
+      "Спорт и фитнес",
+      "Технологии и гаджеты",
+      "Окружающая среда",
+      "Направления для путешествий",
+      "Телесериалы",
+      "Планы на выходные",
+      "Работа и карьерные цели",
+      "Работа из дома"
+    ]
   },
   "voiceRecorder": {
     "record": "Запись",
@@ -980,7 +1012,7 @@
   },
   "whatsNew": {
     "title": "Что нового",
-    "version": "v1.6.6",
+    "version": "v1.6.7",
     "cta": "Понятно",
     "entry1": {
       "label": "Улучшения плана обучения",
