+++ specs/version.md
@@ -1,6 +1,6 @@
 # Version
 
-**1.6.6**
+**1.6.7**
 
 > Canonical project version. Update this file when bumping.
 > Full history in [CHANGELOG.md](../CHANGELOG.md).
